float to double + cerrs

+++ base_class.hpp
@@ -18,9 +18,9 @@ class base_class
         node* node1;                    //Used to tell which nodes this component is connected to
         node* node2;
         string name;       
-        float value;
-        float prev_cv = 0;              //Used for transient sims to store the previous voltage/current value
-	    float tot_acc = 0;
+        double value;
+        double prev_cv = 0;              //Used for transient sims to store the previous voltage/current value
+	    double tot_acc = 0;
         char type;
 
 
@@ -45,25 +45,25 @@ class base_class
 
 
         //Sets the previous value of current/voltage
-        virtual void set_prev_cv(float cv){
+        virtual void set_prev_cv(double cv){
             prev_cv = cv;
         }
         
         //Returns the previous value of current/voltage
-        virtual float return_prev_cv(){
+        virtual double return_prev_cv(){
             return prev_cv;
         }
 
         //Returns current/voltage/resistance depending on subclass (Will be overwritten in the appropriate subclasses)
-        virtual float return_value(float t, bool final_loop_checker){
+        virtual double return_value(double t, bool final_loop_checker){
             return 1;
         }
 
-        virtual float return_tot_acc(){
+        virtual double return_tot_acc(){
             return tot_acc;
         }
 
-        virtual void set_tot_acc(float acc){
+        virtual void set_tot_acc(double acc){
             tot_acc = acc;
         }
