+++ source.hpp
@@ -10,12 +10,12 @@ class source: public base_class
 {
     protected:
         string output_type;
-        float frequency, amplitude;
+        double frequency, amplitude;
 
 
     public:
         //Constructor
-        source(char c_type, string c_output_type, node *c_node1, node *c_node2, string c_name, float c_frequency, float c_value, float c_amplitude){
+        source(char c_type, string c_output_type, node *c_node1, node *c_node2, string c_name, double c_frequency, double c_value, double c_amplitude){
             type = c_type;
             output_type = c_output_type;
             node1 = c_node1;
@@ -28,7 +28,7 @@ class source: public base_class
 
 
         //Returns the output value of the source depending on the type of the source
-        float return_value(float t, bool final_loop_checker){
+        double return_value(double t, bool final_loop_checker){
             if (output_type == "sine"){
                 return ( amplitude*sin(2*M_PI*frequency*t) ) + value;
             }
